Improves build process & fixes minor scripting errors.
- Avoid executing cmake/make/make install if project has already being built and installed. This behaviour is controlled through the `SKIP_INSTALLED` env variable.
- Removes invalid syntax in cmake_install function (it was bash)
- Adds a `generate_env_script.sh` that generates an environment script whose objective is to setup the shell environment to use the rocm installation.
- Installation succeded on a Ubuntu machine.

+++ scripts/common/install_rocm_deps.sh
@@ -25,12 +25,16 @@ autoreconf_build https://gitlab.freedesktop.org/xorg/lib/libxcb/-/archive/libxcb
 autoreconf_build https://gitlab.freedesktop.org/xorg/lib/libpciaccess/-/archive/libpciaccess-0.16/libpciaccess-libpciaccess-0.16.tar.gz
 [ -e kbproto ] || run_command git clone https://gitlab.freedesktop.org/xorg/proto/kbproto
 run_command cd kbproto
-run_command git checkout kbproto-1.0.7
-run_command aclocal
-run_command autoreconf -if
-run_command ./configure --prefix="${INSTALL_DIR}"
-run_command make -j $NCORES install
-
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "kbproto already installed. Skipping.."
+else
+    run_command git checkout kbproto-1.0.7
+    run_command aclocal
+    run_command autoreconf -if
+    run_command ./configure --prefix="${INSTALL_DIR}"
+    run_command make -j $NCORES install
+    run_command touch rfs_installed
+fi
 autoreconf_build https://www.x.org/releases/individual/lib/libX11-1.6.8.tar.gz
 
 # -----------------------------------------------------------------------------------------------
@@ -59,9 +63,13 @@ fi
 
 
 wget_untar_cd https://gitlab.freedesktop.org/mesa/drm/-/archive/libdrm-2.4.111/drm-libdrm-2.4.111.tar.gz
-run_command  meson.py builddir --prefix="${INSTALL_DIR}"
-run_command ninja -C builddir -j $NCORES install
-
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "libdrm already installed. Skipping.."
+else
+    run_command  meson.py builddir --prefix="${INSTALL_DIR}"
+    run_command ninja -C builddir -j $NCORES install
+    run_command touch rfs_installed
+fi
 # -----------------------------------------------------------------------------------------------------
 #                                         sqlite3
 #------------------------------------------------------------------------------------------------------
@@ -72,8 +80,12 @@ configure_build https://www.sqlite.org/snapshot/sqlite-snapshot-202205121156.tar
 #                                          bzip2
 #------------------------------------------------------------------------------------------------------
 wget_untar_cd https://sourceware.org/pub/bzip2/bzip2-1.0.8.tar.gz
-run_command make -j $NCORES install CFLAGS=-fPIC  PREFIX="${INSTALL_DIR}"
-
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "bzip2 already installed. Skipping.."
+else
+    run_command make -j $NCORES install CFLAGS=-fPIC  PREFIX="${INSTALL_DIR}"
+    run_command touch rfs_installed
+fi
 # -----------------------------------------------------------------------------------------------------
 #                                          boost
 #------------------------------------------------------------------------------------------------------
@@ -82,42 +94,62 @@ cd ${BUILD_FOLDER}
 [ -e boost_1_72_0_rc2.tar.gz ] || wget https://boostorg.jfrog.io/artifactory/main/release/1.72.0/source/boost_1_72_0_rc2.tar.gz
 [ -e boost_1_72_0 ] || tar xf boost_1_72_0_rc2.tar.gz
 cd boost_1_72_0
-# the following is needed due to a bug in boost 1.72 installation process
-OLD_CPLUS_VAR=$CPLUS_INCLUDE_PATH
-export CPLUS_INCLUDE_PATH=""
-( unset CPLUS_INCLUDE_PATH; unset CPATH; ./bootstrap.sh --prefix="${INSTALL_DIR}" )
-export CPLUS_INCLUDE_PATH=$OLD_CPLUS_VAR
-run_command ./b2 headers
-run_command ./b2 -j$NCORES cxxflags=-fPIC cflags=-fPIC install toolset=gcc --with=all --prefix="${INSTALL_DIR}"
-
+BOOST_ROOT_DIR=`pwd`
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "Boost already installed. Skipping.."
+else
+    # the following is needed due to a bug in boost 1.72 installation process
+    OLD_CPLUS_VAR=$CPLUS_INCLUDE_PATH
+    export CPLUS_INCLUDE_PATH=""
+    ( unset CPLUS_INCLUDE_PATH; unset CPATH; ./bootstrap.sh --prefix="${INSTALL_DIR}" )
+    export CPLUS_INCLUDE_PATH=$OLD_CPLUS_VAR
+    run_command ./b2 headers
+    run_command ./b2 -j$NCORES cxxflags=-fPIC cflags=-fPIC install toolset=gcc --with=all --prefix="${INSTALL_DIR}"
+    run_command touch "${BOOST_ROOT_DIR}/rfs_installed"
+fi
 # -----------------------------------------------------------------------------------------------------
 #                                          elfutils
 #------------------------------------------------------------------------------------------------------
 cd ${BUILD_FOLDER}
 [ -e elfutils ] || run_command git clone git://sourceware.org/git/elfutils.git
 run_command cd elfutils
-run_command git checkout elfutils-0.187
-run_command autoreconf -i -f
-run_command ./configure --enable-maintainer-mode --disable-libdebuginfod --disable-debuginfod --prefix="${INSTALL_DIR}"
-run_command make -j $NCORES install
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "elfutils already installed. Skipping.."
+else
+    run_command git checkout elfutils-0.187
+    run_command autoreconf -i -f
+    run_command ./configure --enable-maintainer-mode --disable-libdebuginfod --disable-debuginfod --prefix="${INSTALL_DIR}"
+    run_command make -j $NCORES install
+    run_command touch "rfs_installed"
+fi
 
 # -----------------------------------------------------------------------------------------------------
 #                                          gmplib
 #------------------------------------------------------------------------------------------------------
 cd ${BUILD_FOLDER}
-wget https://gmplib.org/download/gmp/gmp-6.2.1.tar.xz
-run_command tar -xf gmp-6.2.1.tar.xz
+[ -e gmp-6.2.1.tar.xz ] || run_command wget https://gmplib.org/download/gmp/gmp-6.2.1.tar.xz
+[ -e gmp-6.2.1 ] || run_command tar -xf gmp-6.2.1.tar.xz
 run_command cd ${BUILD_FOLDER}/gmp-6.2.1
-run_command ./configure --prefix=${INSTALL_DIR}
-run_command make -j $NCORES
-run_command make install
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "gmp already installed. Skipping.."
+else
+    run_command ./configure --prefix=${INSTALL_DIR}
+    run_command make -j $NCORES
+    run_command make install
+    run_command touch "rfs_installed"
+fi
 
 
 # -----------------------------------------------------------------------------------------------------
 #                                          libexpat
 #------------------------------------------------------------------------------------------------------
 wget_untar_cd https://github.com/libexpat/libexpat/releases/download/R_2_4_8/expat-2.4.8.tar.gz
-mkdir build
-cd build
-run_command cmake -DCMAKE_BUILD_TYPE=${BUILD_TYPE} -DCMAKE_INSTALL_PREFIX=${INSTALL_DIR} ..
-run_command make -j $NCORES install
+if [ -e rfs_installed ] &&  [ ${SKIP_INSTALLED} -eq 1 ]; then
+  	echo "gmp already installed. Skipping.."
+else
+    mkdir build
+    cd build
+    run_command cmake -DCMAKE_BUILD_TYPE=${BUILD_TYPE} -DCMAKE_INSTALL_PREFIX=${INSTALL_DIR} ..
+    run_command make -j $NCORES install
+    run_command touch ../rfs_installed
+fi
